split hardhat tests
